requirements.txt
# Core
python-dotenv==1.0.1
requests==2.32.3
pandas==2.2.2

# Embeddings
cohere==5.5.8

# Vector DB
pinecone==4.1.0

# LLM
groq==0.9.0

# Backend (Phase 3)
fastapi==0.111.1
uvicorn[standard]==0.30.1
httpx==0.27.0

# Utils
tqdm==4.66.4
